// types.ts
export enum Sport {
  NBA = 'NBA',
  NFL = 'NFL',
  BASEBALL = 'BASEBALL',
  HOCKEY = 'HOCKEY',
  SOCCER = 'SOCCER',
  COLLEGE_FOOTBALL = 'COLLEGE_FOOTBALL',
  COLLEGE_BASKETBALL = 'COLLEGE_BASKETBALL',
  WNBA = 'WNBA',
  TENNIS = 'TENNIS',
  GOLF = 'GOLF',
  MMA = 'MMA',
  BASKETBALL = 'BASKETBALL'
}
// ...
export interface Linescore {
  value: number;
  label: string;
  period: number;
}

export interface Team {
  id: string;
  name: string;
  shortName: string;
  abbreviation?: string;
  logo: string;
  color?: string;
  record?: string;
  rank?: number;
  score: number;
  linescores?: Linescore[];

  stadiumThumb?: string;
  fanArt?: string;
  stadium?: string;
}

export interface MatchOdds {
  provider?: string;
  hasOdds?: boolean;
  homeWin?: string | number;
  awayWin?: string | number;
  draw?: string | number;
  spread?: string | number;
  homeSpread?: string | number;
  awaySpread?: string | number;
  overUnder?: string | number;
  over?: string | number;
  under?: string | number;
  moneylineHome?: string | number;
  moneylineAway?: string | number;
  totalOver?: string | number;
  homeSpreadOdds?: string | number;
  awaySpreadOdds?: string | number;

  winProbability?: number;
}